requirements.txt
streamlit
python-dotenv
cohere
llama-index
chromadb